pyproject.toml
[project]
name = "dj-dcms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.1.7",
    "django-cleanup==9.0.0",
    "django-environ==0.12.0",
    "django-extensions==3.2.3",
    "django-widget-tweaks==1.5.0",
    "factory-boy==3.3.3",
    "pillow==11.1.0",
    "psycopg==3.2.6",
    "pydantic[email]>=2.11.1",
    "pytest-cov==6.0.0",
    "pytest-django==4.10.0",
    "uvicorn==0.34.0",
]
